Add open_from_stdin_or_file function
This function allow to initialize an ImageIOBase (HDF5ImageIO when
reading from stdin), that can be used to read image informations.

The ImageIOBase can the reused in CmdReader.

CmdReader is templated on PixelType, so if we want to set PixelType according
to type information in file, it is required to read informations using
open_from_stdin_or_file before creating CmdReader.

When reading in stdin, open_from_stdin_or_file read the wole file content.
When reading from a file on the disk, only meta-informations are readen.